init project team members
