+++ team-members/src/App.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+     Hello
+    </div>
+  );
+}
+
+export default App;
